more work on moving endpoints
