+++ frontend/src/lib/services/api/system-api-service.ts
@@ -1,5 +1,6 @@
 import type { DockerInfo } from '$lib/types/docker-info.type';
 import BaseAPIService from './api-service';
+import { environmentStore } from '$lib/stores/environment.store';
 
 export default class SystemAPIService extends BaseAPIService {
 	async pruneAll(options: {
@@ -15,14 +16,17 @@ export default class SystemAPIService extends BaseAPIService {
 	}
 
 	async startAllStoppedContainers() {
-		return this.handleResponse(this.api.post('/system/containers/start-stopped'));
+		const envId = await environmentStore.getCurrentEnvironmentId();
+		return this.handleResponse(this.api.post(`/environments/${envId}/system/containers/start-stopped`));
 	}
 
 	async stopAllContainers() {
-		return this.handleResponse(this.api.post('/system/containers/stop-all'));
+		const envId = await environmentStore.getCurrentEnvironmentId();
+		return this.handleResponse(this.api.post(`/environments/${envId}/system/containers/stop-all`));
 	}
 
 	async getDockerInfo(): Promise<DockerInfo> {
-		return this.handleResponse(this.api.get('/system/docker/info'));
+		const envId = await environmentStore.getCurrentEnvironmentId();
+		return this.handleResponse(this.api.get(`/environments/${envId}/system/docker/info`));
 	}
 }
